+++ CHANGELOG.md
@@ -29,6 +29,8 @@
   from being inherited from an env file (`envFile`). A `null` value now only ignores the variable
   from the system/shell environment, while still allowing an env file to provide a value.
 - Fixed an issue where default `--summary` for `moon ci` was not being respected.
+- Fixed an issue where resolved lockfile versions were not extracted for a toolchain when a
+  range/requirement was defined for the toolchain `version`.
 
 ## 2.5.1